When image too small or canceling crop, remove image selection
If the user cancels the crop then the image is not cropped.  If one or
both of the image's dimensions are too small, the crop window does not
appear and the image is also not cropped.

If the image is not cropped, the user can still save it, resulting an
image of an undesirable ratio being saved.  This change ensures that the
image selection is removed if the user cancels the crop or the image is
too small.